Updated pin mappings to be more dense to match PCB layout.  Tested shooting successfully.
I created a layout in Fritzing (not included here yet) that maps everything out and I adjusted the pin mapping in code to match.  The only divergance now is swapping IR TX with PWM Receiver input as the IR Remote library has some hardcoded constants for pin3 that are not obvious changes.

I tested with an LTX gun and I was able to use an RC transmitter to cause the board to both shoot via throttle channel and be shot by the gun.  The issue with pin 3 was the problem on the first test.